split the client's `types` and `protocol` mods into api crate
This allows the reflector crate to not require lots of network dependencies.

Also, move InternedAttrKey into the protocol module for clarity.
