+++ client-libraries/rust/modality-api/src/lib.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "client")]
+pub mod protocol;
+pub mod types;
+
+pub use types::*;
