+++ client-libraries/rust/modality-api/src/types.rs
@@ -1,10 +1,7 @@
 use std::{cmp::Ordering, ops::Deref, str::FromStr};
 
-use minicbor::{data::Tag, decode, encode, Decode, Decoder, Encode, Encoder};
 use ordered_float::OrderedFloat;
 pub use uuid::Uuid;
-
-use crate::protocol::{TAG_LOGICAL_TIME, TAG_NS, TAG_TIMELINE_ID};
 
 /// They key naming an attribute.
 #[derive(Debug, Clone, Eq, PartialEq, Hash, PartialOrd)]
@@ -37,17 +34,6 @@ impl AsRef<str> for AttrKey {
 impl std::fmt::Display for AttrKey {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
         write!(f, "{}", self.0)
-    }
-}
-
-/// The numeric representation of an `AttrKey` after it has been declared on a connection.
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
-#[repr(transparent)]
-pub struct InternedAttrKey(pub(crate) u32);
-
-impl From<InternedAttrKey> for u32 {
-    fn from(k: InternedAttrKey) -> Self {
-        k.0
     }
 }
 
@@ -114,13 +100,6 @@ impl From<u64> for Nanoseconds {
 impl std::fmt::Display for Nanoseconds {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}ns", self.0)
-    }
-}
-
-impl Encode for Nanoseconds {
-    fn encode<W: encode::Write>(&self, e: &mut Encoder<W>) -> Result<(), encode::Error<W::Error>> {
-        e.tag(TAG_NS)?.u64(self.get_raw())?;
-        Ok(())
     }
 }
 
@@ -192,13 +171,6 @@ impl std::fmt::Display for LogicalTime {
     }
 }
 
-impl Encode for LogicalTime {
-    fn encode<W: encode::Write>(&self, e: &mut Encoder<W>) -> Result<(), encode::Error<W::Error>> {
-        e.tag(TAG_LOGICAL_TIME)?.encode(self.get_raw())?;
-        Ok(())
-    }
-}
-
 impl FromStr for LogicalTime {
     type Err = ();
     fn from_str(s: &str) -> Result<Self, Self::Err> {
@@ -258,26 +230,6 @@ impl std::fmt::Display for TimelineId {
     }
 }
 
-impl Encode for TimelineId {
-    fn encode<W: encode::Write>(&self, e: &mut Encoder<W>) -> Result<(), encode::Error<W::Error>> {
-        e.tag(TAG_TIMELINE_ID)?.bytes(self.get_raw().as_bytes())?;
-        Ok(())
-    }
-}
-
-impl<'b> Decode<'b> for TimelineId {
-    fn decode(d: &mut Decoder<'b>) -> Result<Self, decode::Error> {
-        let t = d.tag()?;
-        if t != TAG_TIMELINE_ID {
-            return Err(decode::Error::Message("Expected TAG_TIMELINE_ID"));
-        }
-
-        Uuid::from_slice(d.bytes()?)
-            .map(Into::into)
-            .map_err(|_uuid_err| decode::Error::Message("Error decoding uuid for TimelineId"))
-    }
-}
-
 /////////////
 // AttrVal //
 /////////////
@@ -306,44 +258,6 @@ impl std::fmt::Display for AttrVal {
             AttrVal::LogicalTime(lt) => lt.fmt(f),
             AttrVal::TimelineId(tid) => tid.fmt(f),
         }
-    }
-}
-
-impl Encode for AttrVal {
-    fn encode<W: encode::Write>(&self, e: &mut Encoder<W>) -> Result<(), encode::Error<W::Error>> {
-        match self {
-            AttrVal::String(s) => {
-                e.str(s.as_str())?;
-            }
-            AttrVal::Integer(i) => {
-                e.i64(*i)?;
-            }
-            AttrVal::BigInt(bi) => {
-                if **bi >= 0i128 {
-                    e.tag(Tag::PosBignum)?.bytes(&bi.to_be_bytes())?;
-                } else {
-                    // this is what the spec says to do. don't ask me.
-                    e.tag(Tag::NegBignum)?.bytes(&((-1 - **bi).to_be_bytes()))?;
-                }
-            }
-            AttrVal::Float(f) => {
-                e.f64(**f)?;
-            }
-            AttrVal::Bool(b) => {
-                e.bool(*b)?;
-            }
-            AttrVal::Timestamp(ns) => {
-                ns.encode(e)?;
-            }
-            AttrVal::LogicalTime(lt) => {
-                lt.encode(e)?;
-            }
-            AttrVal::TimelineId(tid) => {
-                tid.encode(e)?;
-            }
-        }
-
-        Ok(())
     }
 }
